gcc_comp_m2_v4/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4

# Logging and utilities
structlog==24.1.0

# SERVICE-specific dependencies
# OpenAI (primary LLM service)
openai==1.10.0

# Pinecone (vector database)
pinecone-client==3.0.0

# Authentication/Authorization (JWT handling)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Security Testing Tools
bandit==1.7.5
safety==2.3.5
pytest-cov==4.1.0
requests==2.31.0

# OWASP ZAP Python API (for DAST)
python-owasp-zap-v2.4==0.0.21

# DefectDojo API integration
defectdojo-api==1.1.0

# Additional testing utilities
Faker==22.0.0
hypothesis==6.92.0
